+++ viewer/src/actions/actions.js
@@ -3,6 +3,10 @@ const DjVu = window.DjVu;
 
 const Actions = {
 
+    closeHelpWindowAction: () => ({ type: Consts.CLOSE_HELP_WINDOW_ACTION }),
+    
+    showHelpWindowAction: () => ({ type: Consts.SHOW_HELP_WINDOW_ACTION }),
+
     saveDocumentAction: () => ({ type: Consts.SAVE_DOCUMENT_ACTION }),
 
     startFileLoadingAction: () => ({ type: Consts.START_FILE_LOADING_ACTION }),