+++ viewer/src/components/ModalWindow.jsx
@@ -1,51 +1,41 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { connect } from 'react-redux';
 import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import { faTimesCircle } from '@fortawesome/fontawesome-free-regular';
 
-import Actions from '../actions/actions';
-
-class ModalWindow extends React.Component {
+export default class ModalWindow extends React.Component {
 
     static propTypes = {
-        header: PropTypes.string,
-        message: PropTypes.string,
-        type: PropTypes.string,
-        closeModalWindow: PropTypes.func.isRequired
+        additionalClasses: PropTypes.string,
+        isError: PropTypes.bool,
+        isFixedSize: PropTypes.bool,
+        onClose: PropTypes.func.isRequired
     };
 
     render() {
-        const { header, message, closeModalWindow } = this.props;
+        const { onClose, additionalClasses, isError, isFixedSize } = this.props;
 
-        if (!header) {
-            return null;
-        }
-        const isError = this.props.type === 'error';
-        const classes = cx({
-            "modal_window": true,
-            "error": isError
-        });
+        const classes = {
+            modal_window: true,
+            error: isError,
+            fixed_size: isFixedSize
+        };
+
         return (
             <div className="modal_window_wrapper">
-                <div className="dark_layer" onClick={closeModalWindow} />
-                <div className={classes}>
-                    <div className="header">
-                        {isError ? "Error: " + header : header}
-                        <FontAwesomeIcon
-                            className="close_button"
-                            icon={faTimesCircle}
-                            onClick={closeModalWindow}
-                        />
+                <div className="dark_layer" onClick={onClose} />
+                <div className={`${cx(classes)} ${additionalClasses || ''}`}>
+                    <FontAwesomeIcon
+                        className="close_button"
+                        icon={faTimesCircle}
+                        onClick={onClose}
+                    />
+                    <div className="content">
+                        {this.props.children}
                     </div>
-                    <div className="body">{message}</div>
                 </div>
             </div>
         );
     }
 }
-
-export default connect(null, {
-    closeModalWindow: Actions.closeModalWindowAction
-})(ModalWindow);